+++ calcite-adapter/src/test/java/software/amazon/documentdb/jdbc/query/DocumentDbQueryMappingServiceTest.java
@@ -1863,255 +1863,4 @@ public class DocumentDbQueryMappingServiceTest extends DocumentDbFlapDoodleTest 
                 "{\"$project\": {\"nestedIdCollection__id\": \"$_id\", \"_id\": \"$document._id\", \"field1\": \"$document.field1\"}}"),
                 result.getAggregateOperations().get(1));
     }
-
-    @Test
-    @DisplayName("Tests $limit is produced when max rows is passed.")
-    void testMaxRows() throws SQLException {
-        final String query =
-                String.format("SELECT * FROM \"%s\".\"%s\"", DATABASE_NAME, COLLECTION_NAME);
-        final DocumentDbMqlQueryContext result = queryMapper.get(query, 10);
-        Assertions.assertNotNull(result);
-        Assertions.assertEquals(1, result.getColumnMetaData().size());
-        Assertions.assertEquals(2, result.getAggregateOperations().size());
-        Assertions.assertEquals(BsonDocument.parse(
-                "{\"$project\": {\"testCollection__id\": '$_id', \"_id\": 0}}"),
-                result.getAggregateOperations().get(0));
-        Assertions.assertEquals(
-                BsonDocument.parse("{\"$limit\": 10}"), result.getAggregateOperations().get(1));
-    }
-
-    @Test
-    @DisplayName("Tests $limit and $sort are produced when max rows is set.")
-    void testOrderByWithMaxRows() throws SQLException {
-        final String queryWithAscendingSort =
-                String.format(
-                        "SELECT * FROM \"%s\".\"%s\" ORDER BY \"%s\" ASC",
-                        DATABASE_NAME, COLLECTION_NAME + "_array", "field");
-        final DocumentDbMqlQueryContext result = queryMapper.get(queryWithAscendingSort, 10);
-        Assertions.assertNotNull(result);
-        Assertions.assertEquals(5, result.getColumnMetaData().size());
-        Assertions.assertEquals(5, result.getAggregateOperations().size());
-        Assertions.assertEquals(
-                BsonDocument.parse(
-                        "{\"$match\": {\"$or\": ["
-                                + "{\"array.field\": {\"$exists\": true}}, "
-                                + "{\"array.field1\": {\"$exists\": true}}, "
-                                + "{\"array.field2\": {\"$exists\": true}}]}}"),
-                result.getAggregateOperations().get(0));
-        Assertions.assertEquals(
-                BsonDocument.parse(
-                        "{ \"$unwind\": {"
-                                + "\"path\": \"$array\", "
-                                + "\"includeArrayIndex\" : \"array_index_lvl_0\", "
-                                + "\"preserveNullAndEmptyArrays\": true }}"),
-                result.getAggregateOperations().get(1));
-        Assertions.assertEquals(
-                BsonDocument.parse(
-                        "{\"$project\": "
-                                + "{\"testCollection__id\": \"$_id\", "
-                                + "\"array_index_lvl_0\": \"$array_index_lvl_0\", "
-                                + "\"field\": \"$array.field\", "
-                                + "\"field1\": \"$array.field1\", "
-                                + "\"field2\": \"$array.field2\", "
-                                + "\"_id\": 0}}"),
-                result.getAggregateOperations().get(2));
-        Assertions.assertEquals(
-                BsonDocument.parse("{ \"$sort\": {\"field\": 1 } }"),
-                result.getAggregateOperations().get(3));
-        Assertions.assertEquals(
-                BsonDocument.parse(String.format("{\"$limit\": {\"$numberLong\": \"%d\"}}", 10)), result.getAggregateOperations().get(4));
-    }
-
-    @Test
-    @DisplayName("Tests $limit and $sort are produced when max rows is set and the query has a limit.")
-    void testOrderByWithMaxRowsAndLimit() throws SQLException {
-        final String queryWithAscendingSort =
-                String.format(
-                        "SELECT * FROM \"%s\".\"%s\" ORDER BY \"%s\" ASC LIMIT 5",
-                        DATABASE_NAME, COLLECTION_NAME + "_array", "field");
-        final DocumentDbMqlQueryContext result = queryMapper.get(queryWithAscendingSort, 10);
-        Assertions.assertNotNull(result);
-        Assertions.assertEquals(5, result.getColumnMetaData().size());
-        Assertions.assertEquals(6, result.getAggregateOperations().size());
-        Assertions.assertEquals(
-                BsonDocument.parse(
-                        "{\"$match\": {\"$or\": ["
-                                + "{\"array.field\": {\"$exists\": true}}, "
-                                + "{\"array.field1\": {\"$exists\": true}}, "
-                                + "{\"array.field2\": {\"$exists\": true}}]}}"),
-                result.getAggregateOperations().get(0));
-        Assertions.assertEquals(
-                BsonDocument.parse(
-                        "{ \"$unwind\": {"
-                                + "\"path\": \"$array\", "
-                                + "\"includeArrayIndex\" : \"array_index_lvl_0\", "
-                                + "\"preserveNullAndEmptyArrays\": true }}"),
-                result.getAggregateOperations().get(1));
-        Assertions.assertEquals(
-                BsonDocument.parse(
-                        "{\"$project\": "
-                                + "{\"testCollection__id\": \"$_id\", "
-                                + "\"array_index_lvl_0\": \"$array_index_lvl_0\", "
-                                + "\"field\": \"$array.field\", "
-                                + "\"field1\": \"$array.field1\", "
-                                + "\"field2\": \"$array.field2\", "
-                                + "\"_id\": 0}}"),
-                result.getAggregateOperations().get(2));
-        Assertions.assertEquals(
-                BsonDocument.parse("{ \"$sort\": {\"field\": 1 } }"),
-                result.getAggregateOperations().get(3));
-        Assertions.assertEquals(
-                BsonDocument.parse("{\"$limit\": 5}"), result.getAggregateOperations().get(4));
-        Assertions.assertEquals(
-                BsonDocument.parse("{\"$limit\": 10}"), result.getAggregateOperations().get(5));
-    }
-
-    @Test
-    @DisplayName("Tests $limit and $sort are produced when max rows is set and the subquery has a limit.")
-    void testOrderByWithMaxRowsAndInnerLimit() throws SQLException {
-        final String queryWithAscendingSort =
-                String.format(
-                        "SELECT * FROM ( SELECT * FROM \"%s\".\"%s\" LIMIT 20 ) ORDER BY \"%s\" ASC",
-                        DATABASE_NAME, COLLECTION_NAME + "_array", "field");
-        final DocumentDbMqlQueryContext result = queryMapper.get(queryWithAscendingSort, 10);
-        Assertions.assertNotNull(result);
-        Assertions.assertEquals(5, result.getColumnMetaData().size());
-        Assertions.assertEquals(7, result.getAggregateOperations().size());
-        Assertions.assertEquals(
-                BsonDocument.parse(
-                        "{\"$match\": {\"$or\": ["
-                                + "{\"array.field\": {\"$exists\": true}}, "
-                                + "{\"array.field1\": {\"$exists\": true}}, "
-                                + "{\"array.field2\": {\"$exists\": true}}]}}"),
-                result.getAggregateOperations().get(0));
-        Assertions.assertEquals(
-                BsonDocument.parse(
-                        "{ \"$unwind\": {"
-                                + "\"path\": \"$array\", "
-                                + "\"includeArrayIndex\" : \"array_index_lvl_0\", "
-                                + "\"preserveNullAndEmptyArrays\": true }}"),
-                result.getAggregateOperations().get(1));
-        Assertions.assertEquals(
-                BsonDocument.parse(
-                        "{\"$project\": "
-                                + "{\"testCollection__id\": \"$_id\", "
-                                + "\"array_index_lvl_0\": \"$array_index_lvl_0\", "
-                                + "\"field\": \"$array.field\", "
-                                + "\"field1\": \"$array.field1\", "
-                                + "\"field2\": \"$array.field2\", "
-                                + "\"_id\": 0}}"),
-                result.getAggregateOperations().get(2));
-        Assertions.assertEquals(
-                BsonDocument.parse("{\"$limit\": 20}"), result.getAggregateOperations().get(3));
-        Assertions.assertEquals(
-                BsonDocument.parse(
-                        "{\"$project\": "
-                                + "{\"testCollection__id\": \"$testCollection__id\", "
-                                + "\"array_index_lvl_0\": \"$array_index_lvl_0\", "
-                                + "\"field\": \"$field\", "
-                                + "\"field1\": \"$field1\", "
-                                + "\"field2\": \"$field2\", "
-                                + "\"_id\": 0}}"),
-                result.getAggregateOperations().get(4));
-        Assertions.assertEquals(
-                BsonDocument.parse("{ \"$sort\": {\"field\": 1 } }"),
-                result.getAggregateOperations().get(5));
-        Assertions.assertEquals(
-                BsonDocument.parse("{\"$limit\": 10}"), result.getAggregateOperations().get(6));
-    }
-
-    @Test
-    @DisplayName("Tests $limit and $sort are produced when max rows is set and the sub query and the query has limit.")
-    void testOrderByWithMaxRowsAndInnerLimitAndOuterLimit() throws SQLException {
-        final String queryWithAscendingSort =
-                String.format(
-                        "SELECT * FROM  (SELECT * FROM \"%s\".\"%s\" LIMIT 20)  ORDER BY \"%s\" ASC LIMIT 30",
-                        DATABASE_NAME, COLLECTION_NAME + "_array", "field");
-        final DocumentDbMqlQueryContext result = queryMapper.get(queryWithAscendingSort, 10);
-        Assertions.assertNotNull(result);
-        Assertions.assertEquals(5, result.getColumnMetaData().size());
-        Assertions.assertEquals(8, result.getAggregateOperations().size());
-        Assertions.assertEquals(
-                BsonDocument.parse(
-                        "{\"$match\": {\"$or\": ["
-                                + "{\"array.field\": {\"$exists\": true}}, "
-                                + "{\"array.field1\": {\"$exists\": true}}, "
-                                + "{\"array.field2\": {\"$exists\": true}}]}}"),
-                result.getAggregateOperations().get(0));
-        Assertions.assertEquals(
-                BsonDocument.parse(
-                        "{ \"$unwind\": {"
-                                + "\"path\": \"$array\", "
-                                + "\"includeArrayIndex\" : \"array_index_lvl_0\", "
-                                + "\"preserveNullAndEmptyArrays\": true }}"),
-                result.getAggregateOperations().get(1));
-        Assertions.assertEquals(
-                BsonDocument.parse(
-                        "{\"$project\": "
-                                + "{\"testCollection__id\": \"$_id\", "
-                                + "\"array_index_lvl_0\": \"$array_index_lvl_0\", "
-                                + "\"field\": \"$array.field\", "
-                                + "\"field1\": \"$array.field1\", "
-                                + "\"field2\": \"$array.field2\", "
-                                + "\"_id\": 0}}"),
-                result.getAggregateOperations().get(2));
-        Assertions.assertEquals(
-                BsonDocument.parse("{\"$limit\": 20}"), result.getAggregateOperations().get(3));
-        Assertions.assertEquals(
-                BsonDocument.parse(
-                        "{\"$project\": "
-                                + "{\"testCollection__id\": \"$testCollection__id\", "
-                                + "\"array_index_lvl_0\": \"$array_index_lvl_0\", "
-                                + "\"field\": \"$field\", "
-                                + "\"field1\": \"$field1\", "
-                                + "\"field2\": \"$field2\", "
-                                + "\"_id\": 0}}"),
-                result.getAggregateOperations().get(4));
-        Assertions.assertEquals(
-                BsonDocument.parse("{ \"$sort\": {\"field\": 1 } }"),
-                result.getAggregateOperations().get(5));
-        Assertions.assertEquals(
-                BsonDocument.parse("{\"$limit\": 30}"), result.getAggregateOperations().get(6));
-        Assertions.assertEquals(
-                BsonDocument.parse("{\"$limit\": 10}"), result.getAggregateOperations().get(7));
-    }
-
-    @Test
-    @DisplayName("Tests $limit is produced and $sort is omitted when max rows is set and the sub query and the query has  no limit.")
-    void testOrderByWithMaxRowsAndInnerOrderBy() throws SQLException {
-        final String queryWithAscendingSort =
-                String.format(
-                        "SELECT * FROM  (SELECT * FROM \"%s\".\"%s\" ORDER BY \"%s\" ASC) ",
-                        DATABASE_NAME, COLLECTION_NAME + "_array", "field");
-        final DocumentDbMqlQueryContext result = queryMapper.get(queryWithAscendingSort, 10);
-        Assertions.assertNotNull(result);
-        Assertions.assertEquals(5, result.getColumnMetaData().size());
-        Assertions.assertEquals(4, result.getAggregateOperations().size());
-        Assertions.assertEquals(
-                BsonDocument.parse(
-                        "{\"$match\": {\"$or\": ["
-                                + "{\"array.field\": {\"$exists\": true}}, "
-                                + "{\"array.field1\": {\"$exists\": true}}, "
-                                + "{\"array.field2\": {\"$exists\": true}}]}}"),
-                result.getAggregateOperations().get(0));
-        Assertions.assertEquals(
-                BsonDocument.parse(
-                        "{ \"$unwind\": {"
-                                + "\"path\": \"$array\", "
-                                + "\"includeArrayIndex\" : \"array_index_lvl_0\", "
-                                + "\"preserveNullAndEmptyArrays\": true }}"),
-                result.getAggregateOperations().get(1));
-        Assertions.assertEquals(
-                BsonDocument.parse(
-                        "{\"$project\": "
-                                + "{\"testCollection__id\": \"$_id\", "
-                                + "\"array_index_lvl_0\": \"$array_index_lvl_0\", "
-                                + "\"field\": \"$array.field\", "
-                                + "\"field1\": \"$array.field1\", "
-                                + "\"field2\": \"$array.field2\", "
-                                + "\"_id\": 0}}"),
-                result.getAggregateOperations().get(2));
-        Assertions.assertEquals(
-                BsonDocument.parse("{\"$limit\": 10}"), result.getAggregateOperations().get(3));
-    }
 }
